1. Analyze a phone number The tuple analysis should return 3 pieces of data
An indication of whether the number has a New York dialing code ie. 212 as the first 3 digits
An indication of whether the number is fake having 555 as a prefix code in positions 5 to 7 (numbering from 1)
The last 4 digits of the number.

Implemented the (static) method PhoneNumber.IsFake() to detect whether the phone number is fake using the phone number info produced in task 1.